Finished ONLY view of DebateCreate and DebateDetail
