+++ src/components/Debate/debateComments.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { ListGroup, ListGroupItem } from 'react-bootstrap';
+
+const Debate = (props) => {
+    return (
+        <ListGroup>
+            <ListGroupItem>Hi</ListGroupItem>
+        </ListGroup>
+    )
+}
+
+export default Debate
