--- src/components/Debate/debatelist.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const DebateList = (props) => {
-    return (
-        <div>Hi</div>
-    )
-}
-
-export default DebateList
